refactor: move Report backwards-compatibility code into its own file
Makes report.go a little easier to read.
